requirements.txt
streamlit
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
numpy
